+++ _config/config.yml
@@ -1,9 +1,7 @@
 ---
 Name: catalog-manager
-After: 'framework/*','cms/*'
 ---
-
-SiteConfig:
+SilverStripe\SiteConfig\SiteConfig:
   required_permission:
     - SITETREE_VIEW_ALL
     - SITETREE_EDIT_ALL
