+++ composer.json
@@ -1,12 +1,17 @@
 {
     "name": "littlegiant/silverstripe-singlepageadmin",
-    "type": "silverstripe-module",
+    "type": "silverstripe-vendormodule",
     "description": "Single page administration via a LeftAndMain like interface",
     "license": "MIT",
     "authors": [
         {
             "name": "Stevie Mayhew",
             "email": "stevie@stevie.co.nz",
+            "role": "Developer"
+        },
+        {
+            "name": "Ryan Potter",
+            "email": "ryan.potter@littlegiant.co.nz",
             "role": "Developer"
         }
     ],
