+++ src/SinglePageCMSForm.php
@@ -1,29 +1,35 @@
 <?php
 
+namespace LittleGiant\SilverStripe\SinglePageAdmin;
+
+use SilverStripe\Control\Controller;
+use SilverStripe\Forms\Form;
+
 /**
  * Class SinglePageCMSForm
  */
-class SinglePageCMSForm extends CMSForm
+class SinglePageCMSForm extends Form
 {
     /**
      * Route validation error responses through response negotiator,
      * so they return the correct markup as expected by the requesting client.
      */
-    protected function getValidationErrorResponse() {
+    protected function getValidationErrorResponse()
+    {
         $request = $this->getRequest();
         $negotiator = $this->getResponseNegotiator();
 
-        if($request->isAjax() && $negotiator) {
+        if ($request->isAjax() && $negotiator) {
             $this->setupFormErrors();
             $result = $this->customise(
-                array('EditForm' => $this)
+                ['EditForm' => $this]
             )->renderWith(Controller::curr()->getTemplatesWithSuffix('_Content'));
 
-            return $negotiator->respond($request, array(
-                'CurrentForm' => function() use($result) {
+            return $negotiator->respond($request, [
+                'CurrentForm' => function () use ($result) {
                     return $result;
-                }
-            ));
+                },
+            ]);
         } else {
             return parent::getValidationErrorResponse();
         }
